requirements.txt
discord.py[voice]
python-dotenv
PyNaCl
gtts
wavelink>=3.0.0
